+++ integrationtests/self/uni_stream_test.go
@@ -1,7 +1,6 @@
 package self_test
 
 import (
-	"crypto/tls"
 	"fmt"
 	"io/ioutil"
 	"net"
@@ -10,7 +9,6 @@ import (
 	quic "github.com/lucas-clemente/quic-go"
 	"github.com/lucas-clemente/quic-go/integrationtests/tools/testserver"
 	"github.com/lucas-clemente/quic-go/internal/protocol"
-	"github.com/lucas-clemente/quic-go/internal/testdata"
 
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
@@ -28,7 +26,7 @@ var _ = Describe("Unidirectional Streams", func() {
 	BeforeEach(func() {
 		var err error
 		qconf = &quic.Config{Versions: []protocol.VersionNumber{protocol.VersionTLS}}
-		server, err = quic.ListenAddr("localhost:0", testdata.GetTLSConfig(), qconf)
+		server, err = quic.ListenAddr("localhost:0", getTLSConfig(), qconf)
 		Expect(err).ToNot(HaveOccurred())
 		serverAddr = fmt.Sprintf("localhost:%d", server.Addr().(*net.UDPAddr).Port)
 	})
@@ -82,7 +80,7 @@ var _ = Describe("Unidirectional Streams", func() {
 
 		client, err := quic.DialAddr(
 			serverAddr,
-			&tls.Config{RootCAs: testdata.GetRootCA()},
+			getTLSClientConfig(),
 			qconf,
 		)
 		Expect(err).ToNot(HaveOccurred())
@@ -100,7 +98,7 @@ var _ = Describe("Unidirectional Streams", func() {
 
 		client, err := quic.DialAddr(
 			serverAddr,
-			&tls.Config{RootCAs: testdata.GetRootCA()},
+			getTLSClientConfig(),
 			qconf,
 		)
 		Expect(err).ToNot(HaveOccurred())
@@ -126,7 +124,7 @@ var _ = Describe("Unidirectional Streams", func() {
 
 		client, err := quic.DialAddr(
 			serverAddr,
-			&tls.Config{RootCAs: testdata.GetRootCA()},
+			getTLSClientConfig(),
 			qconf,
 		)
 		Expect(err).ToNot(HaveOccurred())